lf-screener-app/requirements.txt
# Requirements for lf-screener-app
tradingview-screener
python-dotenv
PyYAML
